skanky hard drives make the heart eager to commit.. earlier.. and oftener

+++ AtoZ/AZBoxView/AZBox.m
@@ -241,7 +241,7 @@
 	[super setSelected:state];
 	[self setNeedsDisplay:YES];
 }
-- (void) setInset:(float)inset {
+- (void) setInset:(CGFloat)inset {
 	inset_ = inset;
 	[self setNeedsDisplay:YES];
 }
